Added: Program For Calculating Volume Of Cylinder

+++ basic-programs/VolumeOfCylinder.java
@@ -4,9 +4,9 @@ public class VolumeOfCylinder {
     public static void main(String[] args) {
         Scanner sc =new Scanner(System.in);
         System.out.println("Enter the radius of the cylinder: ");
-        int r = sc.nextInt();
+        double r = sc.nextDouble();
         System.out.println("Enter the height of the cylinder: ");
-        int h = sc.nextInt();
+        Double h = sc.nextDouble();
         double volume = 3.14 * r * r *h;
         System.out.println("The volume of the cylinder is "+volume);
         sc.close();
